+++ src/components/App.jsx
@@ -312,6 +312,21 @@ class App extends Component {
 							{this.props.allocations["total_file_size_of_hazard_datasets"]} used
 						</Typography>
 					</Box>
+					<Box className={classes.status}>
+						<LinearProgress
+							variant="determinate"
+							className={classes.customProgressBar}
+							value={
+								((this.props.usage["total_number_of_dfr3"] ?? 0) /
+									this.props.allocations["total_number_of_dfr3"]) *
+								100
+							}
+						/>
+						<Typography className={classes.fontLight} style={{ fontSize: "10px" }}>
+							DFR3 {this.props.usage["total_number_of_dfr3"] ?? 0} of{" "}
+							{this.props.allocations["total_number_of_dfr3"]} used
+						</Typography>
+					</Box>
 					<Divider orientation="horizontal" />
 					<MenuItem
 						className={classes.denseStyle}