+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Read IN-CORE lab quota from service endpoint [#141](https://github.com/IN-CORE/incore-ui/issues/141)
 
+### Added
+- Display DFR3 quota [#137](https://github.com/IN-CORE/incore-ui/issues/137)
+
 
 ## [1.7.0] - 2023-08-16